Testing different types of weapons
Engine:
Added the "EquipmentModel" which is used in the "ActorModel". It's used to get the type of weapons in the hands of the hero

Flounchy:
Added "ChatBox" which has scrolling text
Changed the "Hand" class so I can pass in a list of points for them to follow
Added a base "Weapon" class for my "Sword" and "Spear" (new class) to implement. Currently there is no real point of this, and I want to try and phase it out
The "Player" now takes a left, and right handed weapon which is set outside the class. We assign a stance that is determined by the weapon(s) that the player is holding. Currently trying to move the hands differently depending on their weapon, but struggling. Think I'm going to combine the hands and weapon into a single object (two-handed only), and will be able to rotate, and move them easier